feat_addProductServiceTests

+++ src/main/java/ro/msg/learning/shop/service/ProductService.java
@@ -19,19 +19,23 @@ public class ProductService {
         this.productRepository = productRepository;
     }
 
-    public void create(Product product) {
-        productRepository.save(product);
+    public Product create(Product product) {
+        return productRepository.save(product);
     }
 
-    public void update(Product product) {
-        productRepository.save(product);
+    public Product update(Product product) {
+        return productRepository.save(product);
     }
 
-    public void delete(UUID productId) {
+    public boolean delete(UUID productId) {
         if(productId == null){
             throw new IllegalArgumentException("Category Id is null");
         }
+        if (!productRepository.existsById(productId)) {
+            throw new ProductNotFoundException("Product not found: " + productId);
+        }
         productRepository.deleteById(productId);
+        return true;
     }
 
     public Product readById(UUID productId) {
